lora-bridge/requirements.txt
pyserial==3.5
paho-mqtt==2.1
python-decouple==3.8
structlog==24.4
pytest==8.3
